requirements.txt
# 核心依赖
xlwt==1.3.0
xlrd==1.2.0
PyYAML>=5.4.1

# 可选依赖（用于扩展功能）
openpyxl>=3.0.9
pandas>=1.3.4
numpy>=1.22
